tests reference-binary.c: dont rely on connect to 1.1.1.1
