+++ tests/gdb-tests/tests/test_command_procinfo.py
@@ -1,12 +1,16 @@
+import os
+
 import gdb
 
 import tests
 
-REFERENCE_BINARY = tests.binaries.get("reference-binary.out")
+REFERENCE_BINARY_NET = tests.binaries.get("reference-binary-net.out")
 
 
 def test_command_procinfo(start_binary):
-    start_binary(REFERENCE_BINARY)
+    start_binary(REFERENCE_BINARY_NET)
+
+    os.system("nc -l -p 31337 2>/dev/null 1>&2 &")
 
     bin_path = gdb.execute("pi pwndbg.gdblib.proc.exe", to_string=True).strip("\n")
     pid = gdb.execute("pi pwndbg.gdblib.proc.pid", to_string=True).strip("\n")
@@ -19,7 +23,7 @@ def test_command_procinfo(start_binary):
 
     assert bin_path in res_list[0]
     assert pid in res_list[1]
-    assert "1.1.1.1:80" in result
+    assert "127.0.0.1:31337" in result
 
 
 def test_command_procinfo_before_binary_start():
